\name{moezipfR-package}
\alias{moezipfR-package}
\alias{moezipfR}
\docType{package}
\title{
Marshall-Olkin Extended Zipf Distribution
}
\description{
Statistical utilities for the analysis of data by means of the Marshall-Olkin Extended Zipf distribution. By plotting the probabilities in log-log scale, this two parameter distribution allows a concave as well as a convex behavior of the function at the beginning of the distribution, maintaining the linearity in the tail. The model contains the Zipf model as a particular case.
}
\details{
\tabular{ll}{
Package: \tab moezipfR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-11-09\cr
License: \tab GPL-3\cr
}
}
\author{
Duarte-López, A.,  Casellas, A. and Pérez-Casany, M.

Maintainer: Ariel Duarte-López <aduarte@ac.ucp.edu>

}

\references{
Casellas, A. (2013) \emph{La distribució Zipf Estesa segons la transformació Marshall-Olkin}. Universitat Politécnica de Catalunya \url{http://upcommons.upc.edu/handle/2099.1/18157}

Duarte-López, A., Prat-Pérez, A., & Pérez-Casany, M. (2015, August). \emph{Using the Marshall-Olkin Extended Zipf Distribution in Graph Generation}. In European Conference on Parallel Processing (pp. 493-502). Springer International Publishing.

Güney, Y., Tuaç, Y., & Arslan, O. (2016). \emph{Marshall–Olkin distribution: parameter estimation and application to cancer data}. Journal of Applied Statistics, 1-13.

Pérez-Casany, M. and Casellas, A. (2013) \emph{Marshall-Olkin Extended Zipf Distribution}. arXiv preprint arXiv:1304.4540 \url{http://arxiv.org/pdf/1304.4540.pdf}
}

