% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moezipfR.utils.getDataMatrix.R
\name{moezipfR.utils.getDataMatrix}
\alias{moezipfR.utils.getDataMatrix}
\title{Convert sample vector to frequency matrix.}
\usage{
moezipfR.utils.getDataMatrix(values)
}
\arguments{
\item{values}{Vector of positive integer values.}
}
\value{
The matrix of frequencies associated to the vector \code{values}.
}
\description{
Converts a sequence of values into a matrix of frequencies.
}
\examples{
data <- rmoezipf(100, 2.5, 1.3)
moezipfR.utils.getDataMatrix(data)

}

