% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmoezipf.R
\name{rmoezipf}
\alias{rmoezipf}
\title{Random number generator.}
\usage{
rmoezipf(n, alpha, beta)
}
\arguments{
\item{n}{Number of random numbers to return.}

\item{alpha}{Value of the \eqn{\alpha} parameter (\eqn{\alpha > 1}).}

\item{beta}{Value of the \eqn{\beta} parameter (\eqn{\beta > 0}).}
}
\value{
Vector containing  the n generated numbers.
}
\description{
Generates random numbers from a MOEZipf distribution with
fixed parameters \eqn{\alpha} and \eqn{\beta}.
}
\examples{
rmoezipf(10, 2.5, 1.3)
}
\references{
{ Young, D. S. (2010). \emph{Tolerance: an R package for estimating tolerance intervals}. Journal of Statistical Software, 36(5), 1-39.}
}

